requirements.txt
PyMySQL==1.0.2
drf_yasg==1.20.0
djangorestframework_simplejwt==4.7.1
Django==3.2.3
djangorestframework==3.12.4
